django/requirements.txt
Django==4.2.18
django-auth-ldap==5.1.0
django-bootstrap5==22.2
django-cors-headers==3.13.0
django_extensions==3.2.3
django-filter==23.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.2
djoser==2.3.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
email_validator==2.2.0
gunicorn==22.0.0
pillow==10.3.0
psycopg2-binary==2.9.3
whitenoise==6.9.0
